+++ about.php
@@ -1,0 +1,22 @@
+<?php require_once "functions.php"; ?>
+
+<!DOCTYPE html>
+
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title></title>
+    <style media="screen">
+      .error {
+        border: 1px solid red;
+      }
+    </style>
+  </head>
+
+  <body>
+
+    <?php require_once "registration-form.php"; ?>
+
+    <?php require_once "footer.php"; ?>
+  </body>
+</html>
