+++ html/footer.php
@@ -1,0 +1,4 @@
+<footer>
+  All rights reserved.
+  <?php echo '©' . date('Y'); ?>
+</footer>
